* fix issues when compiling in c++11 mode 	* implement page up/down key handling in CDataBrowser 	* change scrollbar look to use round borders 	* mouse wheel messages are now only dispatched to mouse enabled views 	* make sure CTextLabel does not handle the mouse wheel message 	* small optimization for drawing strings on Mac 	* fix issue that the OpenGL view does not draw on Windows when used on another thread than the main thread 	* fix issue where the VST3Editor does not update the option menu when bound to a vst 3 parameter 	* implement binding a vst 3 parameter to a CTextLabel in VST3Editor 	* replace custom linked list for child views in CViewContainer by a std::list 	* add UTI to CFileExtension 	* fix CCheckBox drawing on Windows when coordinates are not rounded to whole integers 	* fix tooltip timer not stopped when tooltip window was hidden on Cocoa

+++ trunk/vstgui/lib/platform/mac/carbon/hiviewframe.cpp
@@ -650,7 +650,7 @@ bool HIViewFrame::invalidRect (const CRect& rect)
 {
 	if (isWindowComposited (window))
 	{
-		HIRect r = { {rect.left, rect.top}, {rect.getWidth (), rect.getHeight ()} };
+		HIRect r = { {static_cast<CGFloat>(rect.left), static_cast<CGFloat>(rect.top)}, {static_cast<CGFloat>(rect.getWidth ()), static_cast<CGFloat>(rect.getHeight ())} };
 		HIViewSetNeedsDisplayInRect (controlRef, &r, true);
 	}
 	else
@@ -822,7 +822,7 @@ CView::DragResult HIViewFrame::doDrag (IDataPackage* source, const CPoint& offse
 				CGImageRef cgImage = cgBitmap ? cgBitmap->getCGImage () : 0;
 				if (cgImage)
 				{
-					HIPoint imageOffset = { offset.x, offset.y };
+					HIPoint imageOffset = { static_cast<CGFloat>(offset.x), static_cast<CGFloat>(offset.y) };
 					SetDragImageWithCGImage (drag, cgImage, &imageOffset, 0);
 				}
